+++ server/src/calculator.h
@@ -6,6 +6,6 @@
 int add(int a, int b);
 int sub(int a, int b);
 int mult(int a, int b);
-float division(int a, int b);
+float division(float a, float b);
 
 #endif //TCP_CLIENT_CALCULATOR_H
